# docs/openapi.yaml
openapi: 3.1.0
info:
  title: meili-ingest Gateway API
  version: 0.1.0
  summary: Ingest any content into Meilisearch through configurable pipelines.
  description: |
    The gateway is the public entry point of meili-ingest.

    **Two tiers**

    * `POST /ingest` — auto-routing. The gateway detects the MIME type (magic
      bytes → file extension → UTF-8 sniff), selects the best matching pipeline
      (tenant-scoped user pipelines > global user pipelines > built-ins),
      resolves the target index and starts a Temporal workflow.
    * `POST /ingest/pipeline/{name}` — explicit pipeline, full control.

    **Tenant context**

    Every request is scoped to a `MeiliContext` resolved from, in order:
    `X-Meili-Host` / `X-Meili-Api-Key` / `X-Meili-Project-Id` / `X-Meili-Index`
    headers (injected by Meilisearch Cloud's Envoy and trusted only when
    `X-Meili-Envoy-Secret` matches the gateway's `ENVOY_TRUSTED_HEADER`), the
    `?index=` query parameter, an `Authorization: Bearer <key>` header, and
    finally the gateway's `MEILI_URL` / `MEILI_API_KEY` environment variables.
    When none of these yields a host and API key the request fails with `400` —
    unless the pipeline is **pinned**: every `meili_indexer` step names a
    [Meilisearch connection](/concepts/connections), which then supplies the host
    and key, and no context is needed.

    **Errors** are always `{"error": "<message>", "code": "<snake_case>"}`.
  license:
    name: MIT
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
  - {}
...
paths:
...
  /jobs/{job_id}:
    get:
      tags: [Jobs]
      operationId: getJob
      summary: Job status and progress
      description: |
        Combines Temporal `describe` with the workflow's `progress` query. Falls
        back to the control plane's job cache when the workflow is no longer
        known to Temporal (retention expired).
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/XMeiliProjectId"
        - $ref: "#/components/parameters/XMeiliEnvoySecret"
      responses:
        "200":
          description: Current job state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
              example:
                job_id: 550e8400-e29b-41d4-a716-446655440000
                status: running
                current_step: enrich
                progress:
                  status: running
                  current_step: enrich
                  completed_steps: 2
                  total_steps: 4
                  steps:
                    - { step_id: extract, plugin: pdf_extractor, status: succeeded, document_count: 12, branches: 1 }
                    - { step_id: chunk, plugin: chunker, status: succeeded, document_count: 87, branches: 1 }
                    - { step_id: enrich, plugin: llm_enricher, status: running, document_count: 0, branches: 87 }
                  cancel_requested: false
                pipeline_used: my-pdf-with-enrichment
                target_index: contracts
        "404":
          $ref: "#/components/responses/JobNotFound"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
  /jobs/{job_id}/cancel:
    post:
      tags: [Jobs]
      operationId: cancelJob
      summary: Cancel a running job
      description: Sends the `cancel` signal to the workflow. Steps already running finish; no new step is scheduled.
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/XMeiliProjectId"
        - $ref: "#/components/parameters/XMeiliEnvoySecret"
      responses:
        "202":
          description: Cancellation requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
        "404":
          $ref: "#/components/responses/JobNotFound"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Self-hosted fallback: a Meilisearch API key with write access. Used as
        `api_key` when no `X-Meili-Api-Key` header is present. Combine with
        `X-Meili-Host` (or the gateway's `MEILI_URL`) to target an instance.
  parameters:
...
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
...
    XMeiliProjectId:
      name: X-Meili-Project-Id
      in: header
      required: false
      description: Tenant / project identifier (Envoy-injected).
      schema:
        type: string
      example: xxx
...
    XMeiliEnvoySecret:
      name: X-Meili-Envoy-Secret
      in: header
      required: false
      description: |
        Shared secret proving the request came through Envoy. Required for the
        other `X-Meili-*` headers to be honoured when the gateway runs with
        `ENVOY_TRUSTED_HEADER` set; ignored otherwise.
      schema:
        type: string
  responses:
...
    JobNotFound:
      description: Unknown job.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "job 550e8400-e29b-41d4-a716-446655440000 not found", "code": "not_found" }
...
    UpstreamUnavailable:
      description: Control plane or Temporal unreachable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "control plane unreachable", "code": "upstream_error" }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message.
        code:
          type: string
          description: Stable snake_case error code.
          examples: [missing_context, bad_request, forbidden, not_found, payload_too_large, unsupported_media_type, invalid_pipeline, upstream_error, internal_error]
    JobStatus:
      type: string
      description: Status of a job or step.
      enum: [queued, running, succeeded, failed, cancelled]
...
    StepResult:
      type: object
      description: Per-step outcome recorded in the workflow progress.
      required: [step_id, plugin, status]
      properties:
        step_id:
          type: string
        plugin:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        document_count:
          type: integer
          minimum: 0
          default: 0
          description: Documents produced (0 for binary outputs).
        branches:
          type: integer
          minimum: 1
          default: 1
          description: Parallel branches (1 unless fanned out).
        error:
          type: string
          description: Error message when `status` is `failed`.
    WorkflowProgress:
      type: object
      description: Snapshot returned by the workflow's `progress` query.
      required: [status, completed_steps, total_steps]
      properties:
        status:
          $ref: "#/components/schemas/JobStatus"
        current_step:
          type: string
          description: Step currently executing (or last executed).
        completed_steps:
          type: integer
          minimum: 0
        total_steps:
          type: integer
          minimum: 0
        steps:
          type: array
          default: []
          items:
            $ref: "#/components/schemas/StepResult"
        error:
          type: string
          description: Failure message when `status` is `failed`.
        cancel_requested:
          type: boolean
          default: false
    JobStatusResponse:
      type: object
      required: [job_id, status, progress, pipeline_used]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/JobStatus"
        current_step:
          type: [string, "null"]
        progress:
          $ref: "#/components/schemas/WorkflowProgress"
        pipeline_used:
          type: string
        target_index:
          type: [string, "null"]
        error:
          type: [string, "null"]
...